feat(unwind): Extract evaluator/parsing module
